+++ app/Controllers/Pengajar.php
@@ -16,8 +16,14 @@ class Pengajar extends BaseController
         $this->session = session();
     }
 
+    #################################################### VIEW CONTROL #################################################### 
+
     public function index()
     {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+
         $path = $this->db->table('user_sub_menu')
             ->select('path')->getWhere(['title' => 'Data Pengajar'])->getRowArray();
         $data = [
@@ -29,103 +35,11 @@ class Pengajar extends BaseController
         return view('/backend/pengajar/pengajar', $data);
     }
 
-    public function getDataById($id)
-    {
-        $d = $this->model->find($id);
-        $d['mengajar'] = $this->getNamaRtById($id);
-        echo json_encode($d);
-    }
-
-    public function getNamaRtById($id)
-    {
-        $arr = $this->getNamaRT($id);
-        for ($i = 0; $i < count($arr); $i++) {
-            $arr2[$i] = $arr[$i]['nama'];
-        }
-        $tMengajar = implode(", ", $arr2);
-        return $tMengajar;
-    }
-
-    public function getNamaRT($id)
-    {
-        $data = $this->db->table('rumah_tahfid')
-            ->select('rumah_tahfid.nama as nama')
-            ->join('mengajar', 'rumah_tahfid.id = mengajar.id_rt')->getWhere(['mengajar.id_pengajar' => $id])->getResultArray();
-        return $data;
-    }
-
-    public function getNamaRtLain($id)
-    {
-        $data = $this->db->query("SELECT `a`.`id`,`a`.`nama`
-        FROM `rumah_tahfid` `a`
-        WHERE `a`.`id` NOT IN (SELECT `id_rt` FROM `mengajar` WHERE `id_pengajar` = $id)")->getResultArray();
-        return $data;
-    }
-
-    public function hapusData($id)
-    {
-        $this->mengajar->where('id_pengajar', $id)->delete();
-        $this->hapusGambar($id);
-        $this->model->delete($id);
-        $this->session->setFlashdata('success', 'Data berhasil di hapus');
-        return redirect()->to('/pengajar');
-    }
-
-    public function tambahData()
-    {
-        // CONVERT NOMOR HP KE 62
-        $no_hp =  $this->request->getVar('no');
-        if (substr($no_hp, 0, 1) == 0) {
-            $no_hp = "62" . substr($no_hp, 1);
-        }
-
-        $data = [
-            'nama' => $this->request->getVar('nama'),
-            'alamat' => trim($this->request->getVar('alamat')),
-            'no_hp' => $no_hp,
-            'img' => $this->uploadGambar()
-        ];
-        $this->model->insert($data);
-
-        // tambahkan data ke tabel mengajar
-        $this->tambahDataMengajar();
-        // TAMPILKAN FLASH DATA
-        $this->session->setFlashdata('success', 'Data berhasil di tambahkan');
-        return redirect()->to('/pengajar');
-    }
-
-
-    public function tambahDataMengajar()
-    {
-        // AMBIL SEMUA DATA NAMA PENGAJAR
-        $pengajar = $this->model->findColumn('nama');
-        // AMBIL DATA JUMLAH PENGAJAR, INDEX ARRAY
-        $last = count($pengajar) - 1;
-        // AMBIL ROW DARI DATA INPUTAN TERAKHIR PADA TABLE PENGAJAR
-        $id_pengajar = $this->model->where('nama', $pengajar[$last])->first();
-
-        // AMBIL NAMA TEMPAT RUMAH TAHFIDZ DARI INPUTAN
-        $rt = $this->request->getVar('tempat');
-
-        // LOOP UNTUK MENYIMPAN DATA ID PENGAJAR DAN ID RUMAH TAHFIDZ PADA TABEL MENGAJAR 
-        // SEBANYAK DATA TEMPAT MENGAJAR YANG DIINPUT USER 
-        for ($i = 0; $i < count($rt); $i++) {
-            # CONVERT NAMA RT MENJADI ID RT
-            $id_rt = $this->rt->where('nama', $rt[$i])
-                ->first();
-            // SIMPAN ID RUMAH TAHFIDZ DAN ID PENGAJAR DALAM ARRAY 
-            $dataMengajar = [
-                'id_pengajar' => $id_pengajar['id'],
-                'id_rt' => $id_rt['id']
-            ];
-            // INSERT DATA DALAM ARRAY KE TABEL MENGAJAR
-            $this->mengajar->insert($dataMengajar);
-        }
-    }
-
-    // MENAMPILKAN VIEW EDIT
     public function edit($id)
     {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
         $path = $this->db->table('user_sub_menu')
             ->select('path')->getWhere(['title' => 'Data Pengajar'])->getRowArray();
         $data = [
@@ -138,8 +52,49 @@ class Pengajar extends BaseController
         return view('backend/pengajar/edit', $data);
     }
 
+    public function excel()
+    {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+        $pengajar = $this->db->table('rumah_tahfid')
+            ->select('rumah_tahfid.nama as rt, pengajar.*')
+            ->join('mengajar', 'rumah_tahfid.id = mengajar.id_rt')
+            ->join('pengajar', 'mengajar.id_pengajar = pengajar.id')
+            ->get()->getResultArray();
+        $data = [
+            'pengajar' => $this->model->findAll(),
+            'rt' => $this->rt->findColumn('nama'),
+        ];
+
+        return view('/backend/excel/pengajar', $data);
+    }
+
+    public function print()
+    {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+        $pengajar = $this->db->table('rumah_tahfid')
+            ->select('rumah_tahfid.nama as rt, pengajar.*')
+            ->join('mengajar', 'rumah_tahfid.id = mengajar.id_rt')
+            ->join('pengajar', 'mengajar.id_pengajar = pengajar.id')
+            ->get()->getResultArray();
+        $data = [
+            'pengajar' => $this->model->findAll(),
+            'rt' => $this->rt->findColumn('nama'),
+        ];
+
+        return view('/backend/print/pengajar', $data);
+    }
+
+    ################################################### CRUD DATA ################################################### 
+
     public function updateData($id)
     {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
         $img = $this->model->where('id', $id)->findColumn('img');
 
         $data = [
@@ -218,14 +173,91 @@ class Pengajar extends BaseController
         }
     }
 
+    public function hapusData($id)
+    {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+
+        $this->mengajar->where('id_pengajar', $id)->delete();
+        $this->hapusGambar($id);
+        $this->model->delete($id);
+        $this->session->setFlashdata('success', 'Data berhasil di hapus');
+        return redirect()->to('/pengajar');
+    }
+
+    public function tambahData()
+    {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+        // CONVERT NOMOR HP KE 62
+        $no_hp =  $this->request->getVar('no');
+        if (substr($no_hp, 0, 1) == 0) {
+            $no_hp = "62" . substr($no_hp, 1);
+        }
+
+        $data = [
+            'nama' => $this->request->getVar('nama'),
+            'alamat' => trim($this->request->getVar('alamat')),
+            'no_hp' => $no_hp,
+            'img' => $this->uploadGambar()
+        ];
+        $this->model->insert($data);
+
+        // tambahkan data ke tabel mengajar
+        $this->tambahDataMengajar();
+        // TAMPILKAN FLASH DATA
+        $this->session->setFlashdata('success', 'Data berhasil di tambahkan');
+        return redirect()->to('/pengajar');
+    }
+
+    public function tambahDataMengajar()
+    {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+        // AMBIL SEMUA DATA NAMA PENGAJAR
+        $pengajar = $this->model->findColumn('nama');
+        // AMBIL DATA JUMLAH PENGAJAR, INDEX ARRAY
+        $last = count($pengajar) - 1;
+        // AMBIL ROW DARI DATA INPUTAN TERAKHIR PADA TABLE PENGAJAR
+        $id_pengajar = $this->model->where('nama', $pengajar[$last])->first();
+
+        // AMBIL NAMA TEMPAT RUMAH TAHFIDZ DARI INPUTAN
+        $rt = $this->request->getVar('tempat');
+
+        // LOOP UNTUK MENYIMPAN DATA ID PENGAJAR DAN ID RUMAH TAHFIDZ PADA TABEL MENGAJAR 
+        // SEBANYAK DATA TEMPAT MENGAJAR YANG DIINPUT USER 
+        for ($i = 0; $i < count($rt); $i++) {
+            # CONVERT NAMA RT MENJADI ID RT
+            $id_rt = $this->rt->where('nama', $rt[$i])
+                ->first();
+            // SIMPAN ID RUMAH TAHFIDZ DAN ID PENGAJAR DALAM ARRAY 
+            $dataMengajar = [
+                'id_pengajar' => $id_pengajar['id'],
+                'id_rt' => $id_rt['id']
+            ];
+            // INSERT DATA DALAM ARRAY KE TABEL MENGAJAR
+            $this->mengajar->insert($dataMengajar);
+        }
+    }
+
+    ########################################################## GAMBAR ########################################################## 
+
     public function uploadGambar($id = null)
     {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
+
         $file = $this->request->getFile('gambar');
         $img = $this->model->where('id', $id)->findColumn('img');
         if ($file->isValid()) {
             if ($file->getName() != "profile.png") {
                 $namaFile = $file->getRandomName();
                 $file->move('./assets/img/uploads/profile', $namaFile);
+                $this->thumbnail($namaFile);
             } else {
                 $namaFile = "profile.png";
             }
@@ -241,13 +273,60 @@ class Pengajar extends BaseController
 
     public function hapusGambar($id)
     {
+        if (!session()->get('username')) {
+            return redirect()->to('/login');
+        }
         $img = $this->model->where('id', $id)->findColumn('img');
         if ($img[0] != "profile.png") {
             unlink(FCPATH . '/assets/img/uploads/profile/' . $img[0]);
+            if (file_exists(FCPATH . '/assets/img/low/low_' . $img[0])) {
+                unlink(FCPATH . '/assets/img/low/low_' . $img[0]);
+            }
+            if (file_exists(FCPATH . '/assets/img/thumbnail/thumb_' . $img[0])) {
+                unlink(FCPATH . '/assets/img/thumbnail/thumb_' . $img[0]);
+            }
         }
     }
 
+    public function thumbnail($file)
+    {
+        $this->image->withFile('./assets/img/uploads/profile/' . $file)
+            ->fit(250, 250, 'center')
+            ->save('./assets/img/thumbnail/thumb_' . $file);
+    }
 
-    //--------------------------------------------------------------------
+    ########################################################## get data ########################################################## 
 
+    public function getDataById($id)
+    {
+        $d = $this->model->find($id);
+        $d['mengajar'] = $this->getNamaRtById($id);
+        echo json_encode($d);
+    }
+
+    public function getNamaRtById($id)
+    {
+        $arr = $this->getNamaRT($id);
+        for ($i = 0; $i < count($arr); $i++) {
+            $arr2[$i] = $arr[$i]['nama'];
+        }
+        $tMengajar = implode(", ", $arr2);
+        return $tMengajar;
+    }
+
+    public function getNamaRT($id)
+    {
+        $data = $this->db->table('rumah_tahfid')
+            ->select('rumah_tahfid.nama as nama')
+            ->join('mengajar', 'rumah_tahfid.id = mengajar.id_rt')->getWhere(['mengajar.id_pengajar' => $id])->getResultArray();
+        return $data;
+    }
+
+    public function getNamaRtLain($id)
+    {
+        $data = $this->db->query("SELECT `a`.`id`,`a`.`nama`
+        FROM `rumah_tahfid` `a`
+        WHERE `a`.`id` NOT IN (SELECT `id_rt` FROM `mengajar` WHERE `id_pengajar` = $id)")->getResultArray();
+        return $data;
+    }
 }
